pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "scripts"
version = "1.0.0"
requires-python = "==3.8.19"
dependencies = [
    "topaz-em",
    "numpy<2.0",
    "pandas==1.5",
    "click",
    "mrcfile",
    "starfile",
    "pydantic",
    "matplotlib"
]

authors = [
  {name = "Phil Smoot", email = "phil.smoot@chanzuckerberg.com"},
  {name = "Jonathan Schwartz", email = "jonathan.schwartz@czii.org"}
]
description = "Command line tools to Run Topaz on CryoET Datasets."
readme = "README.md"

[project.scripts]
create_parameter = "scripts.parameters_factor:create_parameter_file"

